Move the methods and declarations for `Box` type to `box.go`
